Change `Rand_Next` to `u32`, document `rand.c` a bit more (#819)
* Change Rand_Next to u32, document rand.c a bit more

* Clean up the quotes a bit, add another note

* Format

* — -> -

* Review

* Remove unnecessary casts

* Remove quote, reformat the comments

* Fix new files

* Make docs a bit more consistent and specific

* Format

+++ src/code/z_actor.c
@@ -4782,7 +4782,7 @@ void Actor_SpawnIceEffects(GlobalContext* globalCtx, Actor* actor, Vec3f limbPos
         yaw = Actor_YawToPoint(actor, limbPos);
 
         for (j = 0; j < effectsPerLimb; j++) {
-            randomYaw = (Rand_Next() >> 0x13) + yaw;
+            randomYaw = ((s32)Rand_Next() >> 0x13) + yaw;
 
             velocity.z = Rand_ZeroFloat(5.0f);
